Insert Logs Into Log Store DB
Logs are now inserted into a Redis DB after being processed and parsed through all enabled templates. Each template matches a field; each field is added to a log in memory. That log is then pushed to the Redis DB, with each log being a hash of the fields as values.

Related To: Issue #2, Issue #3, Issue #6