Add Erudite Host-Anomaly Detection
Code has been added to Erudit in order to identify unknown hosts. If it finds a new host in any of the logs as it performs its scans, it will log it as an anomaly.

In an upcoming commit, we will be recording these known hosts to be persistent so that whenever Erudite performs a scan, it can compare to what it saw during past scans to identify what's different. It will also be generating alerts when it identifies unknown hosts as well.

In order to accomedate this functionality, I had to make the matching engine in anatomize.Parser more robust to handle matching indexes and regex groups within those matches. This was in order to properly parse hosts from logs as this can get tricky.

Relates To: Issue #3, Issue #44
Fixes: Issue #26