Decouple Unknown Host Analysis From Main Anomaly Detection Engine Function
Decoupling the unknown host logic from startAnomalyDetection() is the first step in adding the additional src and dst anomaly detection logic noted in Issue #52.

startAnomalyDetection() will now be responsible for calling the functions for each of the different subengines for anomaly detection (host, src, and dst detection right now).

Relates To: Issue #52